Close the serverTransport when stopping a server.
For some reason the upstream implementation only sets a boolean of
interrupted when stopping a server. When the server is blocked on
Accept() setting that boolean doesn't interrupt that call, so the
server would only shut down after it got one more request. While this
is more of a pain during development, it is also easy to fix.